Move Compose-related filters and renderers into the main objects.
This makes them more discoverable, and more consistent with `ScanScopes.composeTestTagScope`.
