+++ radiography/api/radiography.api
@@ -64,6 +64,7 @@ public final class radiography/ViewFilters {
 	public static final fun and (Lradiography/ViewFilter;Lradiography/ViewFilter;)Lradiography/ViewFilter;
 	public static final fun androidViewFilterFor (Ljava/lang/Class;Lkotlin/jvm/functions/Function1;)Lradiography/ViewFilter;
 	public static final fun getNoFilter ()Lradiography/ViewFilter;
+	public static final fun skipComposeTestTagsFilter ([Ljava/lang/String;)Lradiography/ViewFilter;
 	public static final fun skipIdsViewFilter ([I)Lradiography/ViewFilter;
 }
 
@@ -73,6 +74,7 @@ public abstract interface class radiography/ViewStateRenderer {
 
 public final class radiography/ViewStateRenderers {
 	public static final field CheckableRenderer Lradiography/ViewStateRenderer;
+	public static final field ComposeViewRenderer Lradiography/ViewStateRenderer;
 	public static final field DefaultsIncludingPii Ljava/util/List;
 	public static final field DefaultsNoPii Ljava/util/List;
 	public static final field INSTANCE Lradiography/ViewStateRenderers;
@@ -87,17 +89,6 @@ public final class radiography/ViewStateRenderers {
 public final class radiography/ViewsKt {
 	public static final synthetic fun scan (Landroid/view/View;Ljava/util/List;Lradiography/ViewFilter;)Ljava/lang/String;
 	public static synthetic fun scan$default (Landroid/view/View;Ljava/util/List;Lradiography/ViewFilter;ILjava/lang/Object;)Ljava/lang/String;
-}
-
-public final class radiography/compose/ComposableFilters {
-	public static final field INSTANCE Lradiography/compose/ComposableFilters;
-	public static final fun skipLayoutIdsFilter (Lkotlin/jvm/functions/Function1;)Lradiography/ViewFilter;
-	public static final fun skipTestTagsFilter ([Ljava/lang/String;)Lradiography/ViewFilter;
-}
-
-public final class radiography/compose/ComposableRenderers {
-	public static final field ComposeViewRenderer Lradiography/ViewStateRenderer;
-	public static final field INSTANCE Lradiography/compose/ComposableRenderers;
 }
 
 public abstract interface annotation class radiography/compose/ExperimentalRadiographyComposeApi : java/lang/annotation/Annotation {
